refactor(warnings): split text-overlap suppression helpers
Move duplicate extraction, inline fragment, visual marker, line continuation, overlap-area, and geometry helpers into a text-overlap domain folder so the warning detector remains focused on candidate collection and emission.
